refactor(dto.post): sub package
