--- src/main/kotlin/com/quicoment/demo/dto/PostRequest.kt
@@ -1,3 +1,0 @@
-package com.quicoment.demo.dto
-
-data class PostRequest(val title: String?, val content: String?, val password: String?)
