+++ src/main/kotlin/com/quicoment/demo/dto/post/PostRequest.kt
@@ -1,0 +1,3 @@
+package com.quicoment.demo.dto.post
+
+data class PostRequest(val title: String?, val content: String?, val password: String?)
