+++ src/main/kotlin/com/quicoment/demo/dto/post/PostResponse.kt
@@ -1,3 +1,3 @@
-package com.quicoment.demo.dto
+package com.quicoment.demo.dto.post
 
 data class PostResponse(val id: Long?, val title: String, val content: String, val password: String)
